Store data in config_entry.runtime_data
