+++ homeassistant/components/brother/const.py
@@ -2,12 +2,13 @@
 
 from __future__ import annotations
 
+from datetime import timedelta
 from typing import Final
-
-DATA_CONFIG_ENTRY: Final = "config_entry"
 
 DOMAIN: Final = "brother"
 
 PRINTER_TYPES: Final = ["laser", "ink"]
 
 SNMP: Final = "snmp"
+
+UPDATE_INTERVAL = timedelta(seconds=30)
